# core/interaction-api/src/main/resources/main-service-openapi.yaml
openapi: 3.0.1
info:
  description: Documentation "Explore With Me" API v1.0
  title: '"Explore With Me" API сервер'
  version: '1.0'
servers:
  - description: Generated server url
    url: http://localhost:8080
...
paths:
...
  /admin/events:
    get:
      description: >-
        Эндпоинт возвращает полную информацию обо всех событиях подходящих под
        переданные условия


        В случае, если по заданным фильтрам не найдено ни одного события,
        возвращает пустой список
      operationId: getEvents_2
      parameters:
        - description: список id пользователей, чьи события нужно найти
          in: query
          name: users
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - description: список состояний в которых находятся искомые события
          in: query
          name: states
          required: false
          schema:
            type: array
            items:
              type: string
        - description: список id категорий в которых будет вестись поиск
          in: query
          name: categories
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - description: дата и время не раньше которых должно произойти событие
          in: query
          name: rangeStart
          required: false
          schema:
            type: string
        - description: дата и время не позже которых должно произойти событие
          in: query
          name: rangeEnd
          required: false
          schema:
            type: string
        - description: >-
            количество событий, которые нужно пропустить для формирования
            текущего набора
          in: query
          name: from
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - description: количество событий в наборе
          in: query
          name: size
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventFullDto'
          description: События найдены
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: >-
                  Failed to convert value of type java.lang.String to required
                  type int; nested exception is java.lang.NumberFormatException:
                  For input string: ad
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
      summary: Поиск событий
      tags:
        - 'Admin: События'
  /admin/events/{eventId}:
    patch:
      description: >-
        Редактирование данных любого события администратором. Валидация данных
        не требуется.

        Обратите внимание:
         - дата начала изменяемого события должна быть не ранее чем за час от даты публикации. (Ожидается код ошибки 409)
        - событие можно публиковать, только если оно в состоянии ожидания
        публикации (Ожидается код ошибки 409)

        - событие можно отклонить, только если оно еще не опубликовано
        (Ожидается код ошибки 409)
      operationId: updateEvent_1
      parameters:
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventAdminRequest'
        description: Данные для изменения информации о событии
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие отредактировано
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=2 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
        '409':
          content:
            application/json:
              example:
                status: FORBIDDEN
                reason: For the requested operation the conditions are not met.
                message: >-
                  Cannot publish the event because it's not in the right state:
                  PUBLISHED
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не удовлетворяет правилам редактирования
      summary: Редактирование данных события и его статуса (отклонение/публикация).
      tags:
        - 'Admin: События'
...
  /events/{id}:
    get:
      description: >-
        Обратите внимание:

        - событие должно быть опубликовано

        - информация о событии должна включать в себя количество просмотров и
        количество подтвержденных запросов

        - информацию о том, что по этому эндпоинту был осуществлен и обработан
        запрос, нужно сохранить в сервисе статистики


        В случае, если события с заданным id не найдено, возвращает статус код
        404
      operationId: getEvent_1
      parameters:
        - description: id события
          in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие найдено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: >-
                  Failed to convert value of type java.lang.String to required
                  type int; nested exception is java.lang.NumberFormatException:
                  For input string: ad
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=13 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
      summary: >-
        Получение подробной информации об опубликованном событии по его
        идентификатору
      tags:
        - 'Public: События'
...
  /users/{userId}/events/{eventId}:
    get:
      description: >-
        В случае, если события с заданным id не найдено, возвращает статус код
        404
      operationId: getEvent
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие найдено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: >-
                  Failed to convert value of type java.lang.String to required
                  type long; nested exception is
                  java.lang.NumberFormatException: For input string: ad
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=13 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
      summary: Получение полной информации о событии добавленном текущим пользователем
      tags:
        - 'Private: События'
    patch:
      description: >
        Обратите внимание:

        - изменить можно только отмененные события или события в состоянии
        ожидания модерации (Ожидается код ошибки 409)

        - дата и время на которые намечено событие не может быть раньше, чем
        через два часа от текущего момента (Ожидается код ошибки 409)
      operationId: updateEvent
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id редактируемого события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventUserRequest'
        description: Новые данные события
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие обновлено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: Event must not be published
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=283 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
        '409':
          content:
            application/json:
              example:
                status: FORBIDDEN
                reason: For the requested operation the conditions are not met.
                message: Only pending or canceled events can be changed
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не удовлетворяет правилам редактирования
      summary: Изменение события добавленного текущим пользователем
      tags:
        - 'Private: События'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        errors:
          type: array
          description: Список стектрейсов или описания ошибок
          example: []
          items:
            type: string
            description: Список стектрейсов или описания ошибок
            example: '[]'
        message:
          type: string
          description: Сообщение об ошибке
          example: Only pending or canceled events can be changed
        reason:
          type: string
          description: Общее описание причины ошибки
          example: For the requested operation the conditions are not met.
        status:
          type: string
          description: Код статуса HTTP-ответа
          example: FORBIDDEN
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        timestamp:
          type: string
          description: >-
            Дата и время когда произошла ошибка (в формате "yyyy-MM-dd
            HH:mm:ss")
          example: '2022-06-09 06:27:23'
      description: Сведения об ошибке
    CategoryDto:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор категории
          format: int64
          readOnly: true
          example: 1
        name:
          maxLength: 50
          minLength: 1
          type: string
          description: Название категории
          example: Концерты
      description: Категория
...
    EventFullDto:
      required:
        - annotation
        - category
        - eventDate
        - initiator
        - location
        - paid
        - title
      type: object
      properties:
        annotation:
          type: string
          description: Краткое описание
          example: >-
            Эксклюзивность нашего шоу гарантирует привлечение максимальной
            зрительской аудитории
        category:
          $ref: '#/components/schemas/CategoryDto'
        confirmedRequests:
          type: integer
          description: Количество одобренных заявок на участие в данном событии
          format: int64
          example: 5
        createdOn:
          type: string
          description: Дата и время создания события (в формате "yyyy-MM-dd HH:mm:ss")
          example: '2022-09-06 11:00:23'
        description:
          type: string
          description: Полное описание события
          example: >-
            Что получится, если соединить кукурузу и полёт? Создатели "Шоу
            летающей кукурузы" испытали эту идею на практике и воплотили в жизнь
            инновационный проект, предлагающий свежий взгляд на развлечения...
        eventDate:
          type: string
          description: >-
            Дата и время на которые намечено событие (в формате "yyyy-MM-dd
            HH:mm:ss")
          example: '2024-12-31 15:10:05'
        id:
          type: integer
          description: Идентификатор
          format: int64
          example: 1
        initiator:
          $ref: '#/components/schemas/UserShortDto'
        location:
          $ref: '#/components/schemas/Location'
        paid:
          type: boolean
          description: Нужно ли оплачивать участие
          example: true
        participantLimit:
          type: integer
          description: >-
            Ограничение на количество участников. Значение 0 - означает
            отсутствие ограничения
          format: int32
          example: 10
          default: 0
        publishedOn:
          type: string
          description: Дата и время публикации события (в формате "yyyy-MM-dd HH:mm:ss")
          example: '2022-09-06 15:10:05'
        requestModeration:
          type: boolean
          description: Нужна ли пре-модерация заявок на участие
          example: true
          default: true
        state:
          type: string
          description: Список состояний жизненного цикла события
          example: PUBLISHED
          enum:
            - PENDING
            - PUBLISHED
            - CANCELED
        title:
          type: string
          description: Заголовок
          example: Знаменитое шоу 'Летающая кукуруза'
        views:
          type: integer
          description: Количество просмотрев события
          format: int64
          example: 999
...
    Location:
      type: object
      properties:
        lat:
          type: number
          description: Широта
          format: float
          example: 55.754167
        lon:
          type: number
          description: Долгота
          format: float
          example: 37.62
      description: Широта и долгота места проведения события
...
    UpdateEventAdminRequest:
      type: object
      properties:
        annotation:
          maxLength: 2000
          minLength: 20
          type: string
          description: Новая аннотация
          example: >-
            Сап прогулки по рекам и каналам – это возможность увидеть
            Практикбург с другого ракурса
        category:
          type: integer
          description: Новая категория
          format: int64
          example: 3
        description:
          maxLength: 7000
          minLength: 20
          type: string
          description: Новое описание
          example: >-
            От английского SUP - Stand Up Paddle — "стоя на доске с веслом",
            гавайская разновидность сёрфинга, в котором серфер, стоя на доске,
            катается на волнах и при этом гребет веслом, а не руками, как в
            классическом серфинге.
        eventDate:
          type: string
          description: >-
            Новые дата и время на которые намечено событие. Дата и время
            указываются в формате "yyyy-MM-dd HH:mm:ss"
          example: '2023-10-11 23:10:05'
        location:
          $ref: '#/components/schemas/Location'
        paid:
          type: boolean
          description: Новое значение флага о платности мероприятия
          example: true
        participantLimit:
          type: integer
          description: Новый лимит пользователей
          format: int32
          example: 7
        requestModeration:
          type: boolean
          description: Нужна ли пре-модерация заявок на участие
          example: false
        stateAction:
          type: string
          description: Новое состояние события
          enum:
            - PUBLISH_EVENT
            - REJECT_EVENT
        title:
          maxLength: 120
          minLength: 3
          type: string
          description: Новый заголовок
          example: Сап прогулки по рекам и каналам
      description: >-
        Данные для изменения информации о событии. Если поле в запросе не
        указано (равно null) - значит изменение этих данных не треубется.
    UpdateEventUserRequest:
      type: object
      properties:
        annotation:
          maxLength: 2000
          minLength: 20
          type: string
          description: Новая аннотация
          example: >-
            Сап прогулки по рекам и каналам – это возможность увидеть
            Практикбург с другого ракурса
        category:
          type: integer
          description: Новая категория
          format: int64
          example: 3
        description:
          maxLength: 7000
          minLength: 20
          type: string
          description: Новое описание
          example: >-
            От английского SUP - Stand Up Paddle — "стоя на доске с веслом",
            гавайская разновидность сёрфинга, в котором серфер, стоя на доске,
            катается на волнах и при этом гребет веслом, а не руками, как в
            классическом серфинге.
        eventDate:
          type: string
          description: >-
            Новые дата и время на которые намечено событие. Дата и время
            указываются в формате "yyyy-MM-dd HH:mm:ss"
          example: '2023-10-11 23:10:05'
        location:
          $ref: '#/components/schemas/Location'
        paid:
          type: boolean
          description: Новое значение флага о платности мероприятия
          example: true
        participantLimit:
          type: integer
          description: Новый лимит пользователей
          format: int32
          example: 7
        requestModeration:
          type: boolean
          description: Нужна ли пре-модерация заявок на участие
          example: false
        stateAction:
          type: string
          description: Изменение сотояния события
          example: CANCEL_REVIEW
          enum:
            - SEND_TO_REVIEW
            - CANCEL_REVIEW
        title:
          maxLength: 120
          minLength: 3
          type: string
          description: Новый заголовок
          example: Сап прогулки по рекам и каналам
      description: >-
        Данные для изменения информации о событии. Если поле в запросе не
        указано (равно null) - значит изменение этих данных не треубется.
...
    UserShortDto:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор
          format: int64
          example: 3
        name:
          type: string
          description: Имя
          example: Фёдоров Матвей
      description: Пользователь (краткая информация)